common: Improve ctr cleanup
This PR improves the ctr cleanup function as it was seen than when we
are using RunC the current cleanup is not able to delete properly
the containers as we got an error saying that we can't delet
non stopped containers. This PR fixes that issue as it was tested
both on RunC and Kata and is working properly.

Fixes #4995

Signed-off-by: Gabriela Cervantes <gabriela.cervantes.tellez@intel.com>